Move constructMerkleTree to plugin_functions
